+++ changelog.md
@@ -7,6 +7,7 @@
 - Add `DiscordHTTPClient.thread_user_get`.
 - Add `RATELIMIT_GROUPS.thread_user_get`.
 - Add `Client.thread_user_get`.
+- `Emoji.__format__` now returns `.as_emoji` if `code` is given as `''`.
 
 #### Bug Fixes